Cleaned up old files
Removed files and code not in use. Still got some JS in static...
