--- flask/app/static/css/custom.css
@@ -1,340 +1,0 @@
-@import "footer.css";
-body {
-  background-color: #0d0d0d;
-  top: 0;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  position: fixed;
-  overflow-y: scroll;
-}
-
-.btn {
-  background-color: transparent;
-  border: none;
-  outline: none;
-  cursor: pointer;
-  color: #fff;
-  font-size: 18px;
-}
-
-.is-beta-blue {
-  color: #0085ff;
-}
-
-.is-uia-red {
-  color: #C8102E;
-}
-
-a,
-button {
-  transition: 0.5s;
-}
-
-a,
-button:hover {
-  color: #F2F3F5;
-}
-
-.is-beta-black {
-  color: #131E29;
-}
-
-.is-beta-white {
-  color: #ffffff;
-}
-
-.icon-center {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  text-align: center;
-}
-
-.modal {
-  display: none;
-  position: fixed;
-  z-index: 1;
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 100%;
-  overflow: auto;
-  background-color: rgba(0, 0, 0, 0.8);
-}
-
-.modal-content {
-  background-color: #000;
-  color: #ffffff;
-  margin: 15% auto;
-  padding: 20px;
-  border: none;
-  border-radius: 4px;
-  width: 50%;
-}
-
-.close {
-  color: #aaa;
-  float: right;
-  font-size: 28px;
-  font-weight: bold;
-  cursor: pointer;
-}
-
-.close:hover,
-.close:focus {
-  color: black;
-  text-decoration: none;
-  cursor: pointer;
-}
-
-.modal-btn {
-  margin-top: 10px;
-}
-
-/* ==== navbar start ==== */
-
-.navbar {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-  background-color: black;
-  opacity: 0.75;
-}
-
-.navbar-item {
-  opacity: 1;
-}
-
-.navbar-start,
-.navbar-end {
-  display: flex;
-  align-items: center;
-  flex: 1 1 auto;
-}
-
-.navbar-end {
-  justify-content: flex-end;
-}
-
-/* ==== navbar end ==== */
-/* ==== tabs start ==== */
-.tabs {
-  width: 100%;
-}
-
-.tabs .custom-link {
-  color: white;
-}
-
-.tabs .custom-link:hover {
-  color: #0085ff;
-}
-
-/* Style for the active tab */
-.tabs li.is-active .custom-link {
-  background-color: #131E29 !important;
-  color: #0085ff;
-}
-
-.tabs ul {
-  display: flex;
-  justify-content: space-around;
-  width: 100%;
-  padding-left: 0;
-  flex-wrap: nowrap;
-  overflow-x: auto;
-  flex-direction: row;
-  overflow-y: hidden;
-  margin-top: 5rem;
-}
-
-.tabs li {
-  flex: 1 0 0;
-  text-align: center;
-  box-sizing: border-box;
-}
-
-.tabs a {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100%;
-}
-
-
-@media (max-width: 768px) {
-  .tabs ul {
-    flex-direction: column;
-    /* Hide horizontal scrollbar */
-    overflow-x: hidden;
-  }
-
-  .tabs li.tab {
-    /* Make tabs full width when stacked */
-    width: 100%;
-  }
-}
-
-/* ==== tabs end ==== */
-
-/* ===== Markdown custom start ===== */
-
-.content-tab {
-  background-color: #131E29;
-  margin: auto;
-  width: 100%;
-  padding: 1em;
-  box-sizing: border-box;
-  border-radius: 5px;
-  transition: background-color 0.3s;
-}
-
-.custom-markdown ul {
-  list-style-type: none;
-  color: #ffffff;
-  padding-left: 20px;
-  /* Required for absolute positioning within li:before */
-  position: relative;
-
-}
-
-.custom-markdown li {
-  /* Spacing between list items */
-  margin-bottom: 0.5em;
-  /* Required for positioning the pseudo element li:before */
-  position: relative;
-
-}
-
-.custom-markdown li:before {
-  /* Default bullet point */
-  content: '•';
-  position: absolute;
-  left: -1.0em;
-}
-
-.custom-markdown li ul li:before {
-  /* Open circle for nested list items */
-  content: '◦';
-  position: absolute;
-  left: -0.75em;
-}
-
-.custom-markdown strong {
-  color: #ffffff;
-  margin-right: 5px;
-}
-
-.custom-markdown a {
-  text-decoration: underline;
-  font-weight: 500;
-  color: #0085ff;
-}
-
-/* ===== Markdown custom end ===== */
-
-/* ===== Dark/light mode toggle start ===== */
-
-/* Mode Toggle Button*/
-#modeToggle {
-  background-color: #333333;
-  color: #FFFFFF;
-  transition: background-color 0.3s, color 0.3s;
-  margin-right: 1rem;
-  /* Adjust right margin for spacing */
-}
-
-#modeToggle:hover {
-  background-color: white;
-  color: black;
-}
-
-@media screen and (max-width: 768px) {
-
-  .navbar-start,
-  .navbar-end {
-    flex: 0 0 auto;
-  }
-}
-
-/* Light Mode - General Styles */
-
-body.light-mode,
-body.light-mode p,
-body.light-mode h1,
-body.light-mode h2,
-body.light-mode h3,
-body.light-mode h4,
-body.light-mode h5,
-body.light-mode h6,
-body.light-mode ul,
-body.light-mode ol,
-body.light-mode li,
-body.light-mode em,
-body.light-mode strong,
-body.light-mode .custom-link {
-  color: #000000;
-}
-
-body.light-mode {
-  background-color: #ffffff;
-  color: #000000;
-}
-
-body.light-mode hr {
-  border-color: #000000;
-}
-
-/* Style for the active tab */
-body.light-mode .tabs li.is-active .custom-link {
-  background-color: #E8E8E8 !important;
-  color: #0085ff;
-}
-
-body.light-mode .custom-link:hover {
-  background-color: #f0f0f0;
-}
-
-body.light-mode .navbar {
-  background-color: #f0f0f0;
-}
-
-body.light-mode .navbar .navbar-item .admin-link {
-  background-color: #E8E8E8;
-  color: #000;
-  transition: background-color 0.3s, color 0.3s;
-}
-
-body.light-mode .content-tab {
-  background-color: #f2f2f2;
-}
-
-/* Light Mode - Mode Toggle Button */
-body.light-mode #modeToggle {
-  background-color: #E8E8E8;
-  color: #333333;
-  border: 1px solid #d1d1d1;
-}
-
-body.light-mode #modeToggle:hover {
-  background-color: #000000;
-  color: #ffffff;
-}
-
-
-body.light-mode .custom-markdown a {
-  color: red;
-}
-
-body.light-mode .modal-content {
-  background-color: #f2f2f2;
-  color: black;
-}
-
-body.light-mode button:hover {
-  color: #000;
-}
-
-/* ===== Dark/light mode toggle end ===== */
