--- flask/app/static/css/footer.css
@@ -1,60 +1,0 @@
-footer {
-  padding: 4rem 2rem;
-  background-color: #0085ff;
-  color: #fff;
-}
-
-  footer .content {
-    max-width: 1100px;
-    margin: 0 auto;
-    display: flex;
-    justify-content: space-between;
-    flex-direction: row;
-    gap: 2rem;
-  }
-
-  footer .content div {
-    text-align: center;
-  }
-
-  footer .content .contact {
-    text-align: left;
-  }
-
-  footer .content .contact .socials {
-    display: flex;
-    gap: 1rem;
-    font-size: 1.75rem;
-    padding-top: 0.5rem;
-  }
-
-  footer .content .resources {
-    text-align: right;
-  }
-
-  footer .content div img {
-    width: 10rem;
-  }
-
-  /* Collapse into a column when small enough */
-  @media screen and (max-width: 600px) {
-    footer .content {
-      flex-direction: column;
-    }
-
-    footer .content div {
-      width: 100%;
-    }
-
-    footer .content .contact {
-      text-align: center;
-    }
-
-    footer .content .contact .socials {
-      justify-content: center;
-    }
-
-    footer .content .resources {
-      text-align: center;
-    }
-  }
