--- flask/app/static/js/discord.js
@@ -1,1 +1,0 @@
-//hello hello this is \m/
