--- flask/app/templates/components/calendar.html
@@ -1,88 +1,0 @@
-<div>
-  <ul class="calendar">
-    <li class="bedpress">
-      <a href="/">
-        <div class="card">
-          <div class="date">
-            <div class="day">
-              05
-            </div>
-            <div class="month">
-              FEB
-            </div>
-          </div>
-          <div class="desc">
-            <h3>Bedpress - SBL</h3>
-            <ul>
-              <li>14:15 - 20:00</li>
-              <li>UiA Grimstad - A2036</li>
-            </ul>
-          </div>
-        </div>
-      </a>
-    </li>
-    <li class="betadev">
-      <a href="/">
-        <div class="card">
-          <div class="date">
-            <div class="day">
-              05
-            </div>
-            <div class="month">
-              FEB
-            </div>
-          </div>
-          <div class="desc">
-            <h3>GameJam 3</h3>
-            <ul>
-              <li>14:15 - 20:00</li>
-              <li>UiA Grimstad - A2036</li>
-            </ul>
-          </div>
-        </div>
-      </a>
-    </li>
-    <li class="">
-      <a href="/">
-        <div class="card">
-          <div class="date">
-            <div class="day">
-              05
-            </div>
-            <div class="month">
-              FEB
-            </div>
-          </div>
-          <div class="desc">
-            <h3>BetaLan</h3>
-            <ul>
-              <li>14:15 - 20:00</li>
-              <li>UiA Grimstad - A2036</li>
-            </ul>
-          </div>
-        </div>
-      </a>
-    </li>
-    <li class="betasec">
-      <a href="/">
-        <div class="card">
-          <div class="date">
-            <div class="day">
-              05
-            </div>
-            <div class="month">
-              FEB
-            </div>
-          </div>
-          <div class="desc">
-            <h3>CTF</h3>
-            <ul>
-              <li>14:15 - 20:00</li>
-              <li>UiA Grimstad - A2036</li>
-            </ul>
-          </div>
-        </div>
-      </a>
-    </li>
-  </ul>
-</div>
