+++ flask/app/templates/components/header.html
@@ -5,7 +5,7 @@
     </a>
     <nav class="menu" id="navigation-menu">
       <a href="/wip">Interessegrupper</a>
-      <a href="/wip">Arrangementer</A>
+      <a href="/wip">Ny student</A>
       <a href="/wip">Karriere</a>
       <a href="/for-bedrifter">For Bedrifter</a>
     </nav>
