--- flask/app/templates/main.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Markdown Viewer</title>
-</head>
-
-<body>
-  <div>
-    <h1>Markdown Viewer</h1>
-    <div>{{ content|safe }}</div>
-  </div>
-</body>
-
-</html>
