+++ flask/app/templates/markdown/betadev.md
@@ -1,0 +1,3 @@
+# BetaDEV
+
+BetaDEV er interessegruppen for studenter som er glade i development og prosjektutvikling.
