+++ flask/app/templates/markdown/ny-student.md
@@ -1,0 +1,28 @@
+# Velkommen til UiA og BETA!
+
+BETA er linjeforeningen for alle datastudentene på UiA Grimstad. Det vil si, om du studerer dataingeniør, cybersikkerhet, eller kunstig intelligens i Grimstad, så er du automatisk medlem! Linjeforeningen er drevet av studentene våres og vi tilrettelegger for at du skal ha et godt sosialt og faglig miljø på campus!
+
+## Hva gjør vi?
+
+Vi arrangerer blant annet:
+
+- **Bedriftspresentasjoner**
+- **BetaLAN**
+- **Capture The Flag**
+- **Fagkvelder**
+- **Game jam**
+- **Sosiale eventer**
+
+Info om alt som skjer postes på [Discord-serveren våres](https://discord.gg/U87D3kPdJ3), så husk å bli med!
+
+## Interessegrupper
+
+BETA har flere interessegrupper og komitéer hvor det er mulig å engasjere seg. 
+
+- **BetaDEV**: Få ekte arbeidserfaringer ved å jobbe med prosjekter sammen med andre, utforsk og diskuter ulike rammeverk og teknologier.
+- **BetaSEC**: Få en praktisk kompetanse i nettverkssikkerhet, arrangere CTF og workshops, diskutere og løse problemer sammen.
+- **BetaLAN**: Arrangere et stort LAN-party hvert semester.
+- **Eventkom**: Arrangere flere sosiale eventer som paint ball, bowling, brettspill, etc.
+- **Bedkom**: Arrangere bedriftspresentasjoner, koble næringslivet med studentene.
+
+
