--- flask/app/templates/month.html
@@ -1,7 +1,0 @@
-<!doctype html>
-<title> game jam entires </title>
-
-<nav>
-  <h1>hello man</h1>
-</nav>
-<section class="content"></section>
