+++ flask/app/templates/placeholder.html
@@ -1,14 +1,22 @@
 {% extends "components/base.html" %}
 
-{% block title %}TITTEL{% endblock %}
+{% block title %}WIP | BETA{% endblock %}
 
 {% block content %}
 
-<div class="construction">
-    <img src="{{ url_for('static', filename='img/construction.png') }}" alt="under-construction, https://commons.wikimedia.org/wiki/File:Under_construction_blue.svg" />
-    <h1>Under arbeid</h1>
-    <p>Har du lyst til å lære eller hjelpe? Vi har åpent møte hver onsdag, så bare kom for en prat!</p>
-    <a href="https://discord.gg/v8b6ZrQ"><button>Kontakt BetaDev</button></a>
+<div class="container container--full container---vertical-centered">
+    <div class="container__content container__content--centered">
+
+    <img class="content__logo" src="{{ url_for('static', filename='img/construction.png') }}" alt="under-construction, https://commons.wikimedia.org/wiki/File:Under_construction_blue.svg" />
+      <h2>Under arbeid</h2>
+      <p>Har du lyst til å lære eller hjelpe? Vi har åpent møte hver onsdag, så bare kom for en prat!</p>
+      <div class="content__btn-menu">
+        <a class="btn-menu__btn content__btn" href="https://discord.gg/v8b6ZrQ">Kontakt BetaDev</a>
+      </div>
+      <span class="wip">Under konstruksjon, sommerprosjekt 2024</span>
+
+    </div>
 </div>
+
 
 {% endblock %}
