--- flask/app/templates/proto.html
@@ -1,21 +1,0 @@
-{% extends "components/base.html" %}
-
-{% block title %}Hjem | BETA{% endblock %}
-
-{% block content %}
-
-<div class="container container--full container---vertical-centered">
-    <div class="container__content container__content--centered">
-
-      <img class="content__logo" src="{{ url_for('static', filename='img/Beta_blue_1.png') }}" alt="under-construction, https://commons.wikimedia.org/wiki/File:Under_construction_blue.svg" />
-      <h2>Linjeforeningen for datastudenter på UiA Grimstad</h2>
-      <div class="content__btn-menu">
-        <a class="btn-menu__btn content__btn" href="/om">Om BETA</a>
-        <a class="btn-menu__btn content__btn" href="https://discord.gg/U87D3kPdJ3" class="btn">Discord</a>
-      </div>
-      <span class="wip">Under konstruksjon, sommerprosjekt 2024</span>
-
-    </div>
-</div>
-
-{% endblock %}
